Split logic and dom
